heic2jpg/Heic-Convert-App/requirements.txt
Pillow
pillow-heif
piexif